Add test for insertUploadVideo, naming fixes
